requirements.txt
asgiref==3.4.0
autopep8==1.5.7
Django==3.2.4
django-dev==0.2.1
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
